FEATURE: Adding Weather Util which populates long term weather db with latest 5 day forecast for each city with an IATA airport in flights db & where city name matches an openweatherapi location
