--- utils/database_utils/add-airports-to-table/README.md
@@ -1,21 +1,0 @@
-
-
-# Current
-https://raw.githubusercontent.com/mborsetti/airportsdata/main/airportsdata/airports.csv
-- Correctly lists glasgow internaltion as glasgow
-- correctly lists glasgow prestwick as glasgow
-- Doesnt have utf issues
-
-
-
-# ourairports.com
-airports.csv in other_data_sources from https://davidmegginson.github.io/ourairports-data/airports.csv
-Can also be found on ourairports.com 
-
-Problematic because it lists "glasgow" as having municipality "Paisley, renfrew", which would send skyscanner and airbnb etc to paisley instead of glasgow
-
-
-# detailed_json
-Cant remeber where i got it from. 
-- municipality data generally ok, 
-- full of utf encoding issues
