--- utils/database_utils/add-flights-to-table/aerodatabox-sql/.gitignore
@@ -1,4 +1,0 @@
-*.db
-*.exe
-aerodatabox-sql
-
