--- utils/database_utils/add-flights-to-table/aerodatabox-sql/fetch-these-flights.yaml
@@ -1,29 +1,0 @@
-flights:
-  - direction: "Departure"
-    airport: "BER"
-    startDate: "03-04-2024"
-    endDate: "06-04-2024"
-  - direction: "Arrival"
-    airport: "BER"
-    startDate: "07-04-2024"
-    endDate: "10-04-2024"
-  - direction: "Departure"
-    airport: "GLA"
-    startDate: "03-04-2024"
-    endDate: "06-04-2024"
-  - direction: "Arrival"
-    airport: "GLA"
-    startDate: "07-04-2024"
-    endDate: "10-04-2024"
-  - direction: "Departure"
-    airport: "EDI"
-    startDate: "03-04-2024"
-    endDate: "06-04-2024"
-  - direction: "Arrival"
-    airport: "EDI"
-    startDate: "07-04-2024"
-    endDate: "10-04-2024"
-
-
-      # You can add more entries with different airports, directions, and date ranges.
-
