+++ utils/database_utils/flights/add-airports-to-table/go.mod
@@ -1,0 +1,8 @@
+module add-airports-to-table
+
+go 1.18
+
+require (
+	github.com/mattn/go-sqlite3 v1.14.22 // indirect
+	golang.org/x/text v0.14.0 // indirect
+)
