+++ utils/database_utils/flights/add-flights-to-table/aerodatabox-sql/.gitignore
@@ -1,0 +1,4 @@
+*.db
+*.exe
+aerodatabox-sql
+
