+++ utils/database_utils/weather/longterm_db/update_weather_db/go.mod
@@ -1,0 +1,9 @@
+module update-weather-db
+
+go 1.18
+
+require (
+	github.com/Tris20/FairFareFinder v0.0.0-20240401193606-23dfe50a010e // indirect
+	github.com/mattn/go-sqlite3 v1.14.22 // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
+)
